Program to generate the filename in the format Openpose application is generating JSON output files name
With every frame application is generating a new JSON file so to read that we need to read the new filename each time.
Current format is : "nnnnnnnnnnn_keypoints.json" where n.. are numbers and start as 000000000000, 000000000001...
Challenge here is to numbers have 0 appended in the filename so with updating of file numbers with frame we may need to change the number of 0 appended.
So for frame 1: 000000000001_keypoints.json [It have 11 zero appended]
while for frame 11: 000000000011_keypoints.json [It have 10 zero appended]
We need to adjust the count of zero to be appended as well